break choice card components out into their files
